# packages/contracts/toolbox/openapi.json
{
  "components": {
    "schemas": {
      "ArtworkRequest": {
        "additionalProperties": false,
        "properties": {
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "quality": {
            "default": 90,
            "maximum": 100,
            "minimum": 1,
            "title": "Quality",
            "type": "integer"
          },
          "size": {
            "default": 1200,
            "maximum": 4000,
            "minimum": 16,
            "title": "Size",
            "type": "integer"
          },
          "square": {
            "default": true,
            "title": "Square",
            "type": "boolean"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "ArtworkRequest",
        "type": "object"
      },
      "ArtworkResult": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "data_base64": {
            "title": "Data Base64",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "mime": {
            "const": "image/jpeg",
            "default": "image/jpeg",
            "title": "Mime",
            "type": "string"
          },
          "source": {
            "enum": [
              "url",
              "path",
              "fixture"
            ],
            "title": "Source",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "width",
          "height",
          "bytes",
          "data_base64",
          "source"
        ],
        "title": "ArtworkResult",
        "type": "object"
      },
      "CookiesTestRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Netscape cookies.txt, inline.",
            "title": "Content"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          }
        },
        "title": "CookiesTestRequest",
        "type": "object"
      },
      "CookiesTestResult": {
        "properties": {
          "authenticated": {
            "description": "A YouTube session cookie is present.",
            "title": "Authenticated",
            "type": "boolean"
          },
          "cookies": {
            "title": "Cookies",
            "type": "integer"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "title": "Domains",
            "type": "array"
          },
          "expired": {
            "default": 0,
            "title": "Expired",
            "type": "integer"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Earliest expiry, ISO-8601 UTC.",
            "title": "Expires At"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "title": "Problems",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "cookies",
          "domains",
          "authenticated"
        ],
        "title": "CookiesTestResult",
        "type": "object"
      },
      "DownloadRequest": {
        "additionalProperties": false,
        "properties": {
          "cookies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path to a Netscape cookies.txt readable by the container.",
            "title": "Cookies"
          },
          "dest_dir": {
            "title": "Dest Dir",
            "type": "string"
          },
          "extra_args": {
            "additionalProperties": true,
            "description": "Raw yt-dlp options merged last. Escape hatch; prefer the named fields.",
            "title": "Extra Args",
            "type": "object"
          },
          "format": {
            "default": "bestaudio",
            "title": "Format",
            "type": "string"
          },
          "id": {
            "description": "Opaque id echoed in every event; used as the file stem.",
            "title": "Id",
            "type": "string"
          },
          "player_client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "yt-dlp `extractor_args.youtube.player_client`, e.g. 'android' or 'web'.",
            "title": "Player Client"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "dest_dir",
          "id"
        ],
        "title": "DownloadRequest",
        "type": "object"
      },
      "ErrorBody": {
        "description": "The JSON body of every failed request, and of the NDJSON ``error`` event.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "hint": {
            "title": "Hint",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "hint",
          "action"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ExtractRequest": {
        "additionalProperties": false,
        "properties": {
          "cookies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path to a Netscape cookies.txt readable by the container.",
            "title": "Cookies"
          },
          "extra_args": {
            "additionalProperties": true,
            "description": "Raw yt-dlp options merged last. Escape hatch; prefer the named fields.",
            "title": "Extra Args",
            "type": "object"
          },
          "player_client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "yt-dlp `extractor_args.youtube.player_client`, e.g. 'android' or 'web'.",
            "title": "Player Client"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ExtractRequest",
        "type": "object"
      },
...
      "FingerprintRequest": {
        "additionalProperties": false,
        "properties": {
          "acoustid_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acoustid Key"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "FingerprintRequest",
        "type": "object"
      },
...
      "ProbeRequest": {
        "additionalProperties": false,
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "ProbeRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Stateless media operations: yt-dlp, mutagen, fpcalc/AcoustID, rsgain.",
    "title": "Music Manager toolbox",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/artwork/prepare": {
      "post": {
        "description": "Crop to square, resize and re-encode a cover as JPEG.",
        "operationId": "prepareArtwork",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtworkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Prepare Artwork",
        "tags": [
          "files"
        ]
      }
    },
    "/cookies/test": {
      "post": {
        "description": "Parse a cookies.txt offline and report whether it is a usable YouTube session.",
        "operationId": "testCookies",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CookiesTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookiesTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Test Cookies",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/download": {
      "post": {
        "description": "Download one file, streaming NDJSON events. A second concurrent call gets 409 LOCKED.",
        "operationId": "download",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/x-ndjson": {}
            },
            "description": "NDJSON stream of progress / postprocess / done / error events."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Download",
        "tags": [
          "download"
        ]
      }
    },
...
    "/extract": {
      "post": {
        "description": "Resolve a URL to its entries without downloading anything.",
        "operationId": "extract",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Extract",
        "tags": [
          "download"
        ]
      }
    },
    "/fingerprint": {
      "post": {
        "description": "Chromaprint fingerprint, plus AcoustID candidates when a key is available.",
        "operationId": "fingerprint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FingerprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FingerprintResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Fingerprint",
        "tags": [
          "files"
        ]
      }
    },
...
    "/place": {
      "post": {
        "description": "Move a file into the library, atomically, creating the folders it needs.",
        "operationId": "place",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Place",
        "tags": [
          "files"
        ]
      }
    },
    "/probe": {
      "post": {
        "description": "Everything ffprobe knows about a file, including every tag present.",
        "operationId": "probe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProbeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProbeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Probe",
        "tags": [
          "files"
        ]
      }
    },
    "/replaygain": {
      "post": {
        "description": "Scan with rsgain and write REPLAYGAIN_* (and R128_* on Opus).",
        "operationId": "replaygain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayGainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayGainResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Replaygain",
        "tags": [
          "files"
        ]
      }
    },
    "/tag": {
      "post": {
        "description": "Write already-projected key/value pairs, pictures and lyrics, then read them back.",
        "operationId": "tag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Tag",
        "tags": [
          "files"
        ]
      }
    },
    "/ytdlp/selftest": {
      "post": {
        "description": "Prove the downloader works: module, binaries, and optionally a real extraction.",
        "operationId": "selftestYtDlp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SelfTestRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelfTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Selftest Ytdlp",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/ytdlp/update": {
      "post": {
        "description": "Upgrade yt-dlp in place and report what changed.",
        "operationId": "updateYtDlp",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Update Ytdlp",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/ytmusic/search": {
      "post": {
        "description": "Find the YouTube Music album playlist (`OLAK5uy_…`) or video behind a release.",
        "operationId": "searchYtMusic",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YtMusicSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YtMusicSearchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A catalogued toolbox error."
          }
        },
        "summary": "Search Ytmusic",
        "tags": [
          "sources"
        ]
      }
    }
  }
}